+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Hide shelf when there are no products to show.
+
 ## [0.6.0] - 2020-12-02
 
 ### Added